poprawki

+++ GACKO.DB/DaoModels/DaoVirtualAccount.cs
@@ -7,9 +7,9 @@ namespace GACKO.DB.DaoModels
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Balance { get; set; }
-        public float Limit { get; set; }
-        public float NotificationBalance { get; set; }
+        public double Balance { get; set; }
+        public double Limit { get; set; }
+        public double NotificationBalance { get; set; }
         public int BankAccountId { get; set; }
         [ForeignKey("BankAccountId")]
         public virtual DaoBankAccount BankAccount { get; set; }
